+++ src/ViewTrait.php
@@ -75,7 +75,7 @@ trait ViewTrait
      * corresponding supported file extensions.
      *
      * ```php
-     * $view = $view->withRenderers(['twig' => new \Yiisoft\View\Twig\ViewRenderer($environment)]);
+     * $view = $view->withRenderers(['twig' => new \Yiisoft\View\Twig\TemplateRenderer($environment)]);
      * ```
      *
      * If no renderer is available for the given view file, the view file will be treated as a normal PHP