+++ src/ViewInterface.php
@@ -50,19 +50,11 @@ interface ViewInterface
     /**
      * Returns a new instance with the specified default view file extension.
      *
-     * @param string $defaultExtension The default view file extension. Default is "php".
+     * @param string $defaultExtension The default view file extension. Default is {@see PHP_EXTENSION}.
      * This will be appended to view file names if they don't have file extensions.
-     * @deprecated Since 8.0.1 and will be removed in the next major version. Use {@see withFallbackExtension()} instead.
+     * @deprecated Since 8.0.1 and will be removed in the next major version.
      */
     public function withDefaultExtension(string $defaultExtension): static;
-
-    /**
-     * Returns a new instance with the specified fallback view file extension.
-     *
-     * @param string $fallbackExtension The fallback view file extension. Default is {@see PHP_EXTENSION}.
-     * This will be appended to view file names if they don't exist.
-     */
-    public function withFallbackExtension(string $fallbackExtension): static;
 
     /**
      * Returns a new instance with the specified view context instance.
@@ -115,16 +107,9 @@ interface ViewInterface
      * Gets the default view file extension.
      *
      * @return string The default view file extension.
-     * @deprecated Since 8.0.1 and will be removed in the next major version. Use {@see withFallbackExtension()} instead.
+     * @deprecated Since 8.0.1 and will be removed in the next major version.
      */
     public function getDefaultExtension(): string;
-
-    /**
-     * Gets the fallback view file extension.
-     *
-     * @return string The fallback view file extension.
-     */
-    public function getFallbackExtension(): string;
 
     /**
      * Gets the theme instance, or `null` if no theme has been set.