+++ CHANGELOG.md
@@ -11,6 +11,7 @@
   `withFallbackExtension()` and `getFallbackExtensions()` (@rustamwin)
 - Bug #232: Fix render templates that contain dots in their name (@rustamwin)
 - New #242: Add `View::getLocale()` and `WebView::getLocale()` methods (@Tigrov)
+- Enh #250: Make event dispatcher in `View` and `WebView` optional (@vjik)
 
 ## 8.0.0 February 16, 2023